Corrigiendo Systems

+++ examples/MatchOne/src/main/java/com/ilargia/games/EntityIndexExtension.java
@@ -1,6 +1,5 @@
 package com.ilargia.games;
 
-import com.badlogic.gdx.utils.ObjectSet;
 import com.ilargia.games.components.Position;
 import com.ilargia.games.core.Entitas;
 import com.ilargia.games.core.GameContext;
@@ -17,19 +16,19 @@ public class EntityIndexExtension {
     final int shiftX = 8;
 
     public static void addEntityIndices(Entitas contexts) {
-        EntityIndex<GameEntity,Integer> positionIndex = new EntityIndex(
+        EntityIndex<GameEntity, Integer> positionIndex = new EntityIndex(
                 contexts.game.getGroup(GameMatcher.Position()),
                 (e, c) -> {
                     Position positionComponent = (Position) c;
                     return positionComponent != null
                             ? positionComponent.x + "," + positionComponent.y
-                            : ((GameEntity)e).getPosition().x + "," + ((GameEntity)e).getPosition().y;
+                            : ((GameEntity) e).getPosition().x + "," + ((GameEntity) e).getPosition().y;
                 }
         );
         contexts.game.addEntityIndex(PositionKey, positionIndex);
     }
 
-    public static Set<GameEntity> getEntitiesWithPosition(GameContext context, int x, int y) {
+    public static Set<GameEntity> getEntitiesWithPosition(GameContext context, float x, float y) {
         EntityIndex<GameEntity, String> index = (EntityIndex<GameEntity, String>) context.getEntityIndex(PositionKey);
         return index.getEntities(x + "," + y);
     }
