+++ examples/MatchOne/src/main/java/com/ilargia/games/components/Position.java
@@ -1,14 +1,14 @@
 package com.ilargia.games.components;
 
-import com.ilargia.games.entitas.codeGenerator.Component;
 import com.ilargia.games.entitas.api.IComponent;
+import com.ilargia.games.entitas.codeGenerator.Component;
 
 @Component(pools = {"Game"})
 public class Position implements IComponent {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
-    public Position(int x, int y) {
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
